Fix page load flicker by reordering JS and CSS
Moving the CSS above the JS avoids the page load flicker. However
the upstream layout doesn't allow overriding this so I've had to
pull in the whole layout in order to make the modifications.

Ideally this should be fixed upstream but the PR to resolve it
has been stuck for some time.

readthedocs/sphinx_rtd_theme#853

If this PR is ever merged we should revert back to using the
inheritance.