Fix missing colons in dask/dask
This fix updates the `layout.html` with changes I grabbed from
readthedocs/sphinx_rtd_theme#838 that relate to the version of the
`htmlwriter` that is used.

This should resolve dask/dask#6696 once a new release of the theme is
cut (I've bumped the version in `__init__.py`) and the update is pulled
in there.